Resolve bpf issue 515

+++ driver/modern_bpf/definitions/events_dimensions.h
@@ -103,7 +103,7 @@
 #define RENAMEAT2_E_SIZE HEADER_LEN
 #define PIPE_E_SIZE HEADER_LEN
 #define PIPE_X_SIZE HEADER_LEN + sizeof(int64_t) * 3 + sizeof(uint64_t) + PARAM_LEN * 4
-#define BPF_E_SIZE HEADER_LEN + sizeof(int64_t) + PARAM_LEN
+#define BPF_E_SIZE HEADER_LEN + sizeof(int32_t) + PARAM_LEN
 #define BPF_X_SIZE HEADER_LEN + sizeof(int64_t) + PARAM_LEN
 #define FLOCK_E_SIZE HEADER_LEN + sizeof(int64_t) + sizeof(uint32_t) + PARAM_LEN * 2
 #define FLOCK_X_SIZE HEADER_LEN + sizeof(int64_t) + PARAM_LEN
